requirements.txt
boto3==1.21.32
botocore==1.24.32
fastapi==0.79.0
python-dotenv==0.20.0
uvicorn==0.18.2
